feat: 策略模式优化折扣计算

+++ group-buy-market-infrastructure/src/main/java/com/hongrui/infrastructure/dao/po/GroupBuyDiscount.java
@@ -17,39 +17,47 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GroupBuyDiscount {
+
     /**
      * 自增ID
      */
     private Long id;
+
     /**
-     * 折扣活动呢ID
+     * 折扣ID
      */
     private Integer discountId;
+
     /**
      * 折扣标题
-     *
      */
     private String discountName;
+
     /**
      * 折扣描述
      */
     private String discountDesc;
+
     /**
-     * 折扣类型（0：base、1：tag）
+     * 折扣类型（0:base、1:tag）
      */
-    private Byte discountType;
+    private Integer discountType;
+
     /**
-     * 营销优惠计划（ZJ：直减、MJ：满减、N元购）
+     * 营销优惠计划（ZJ:直减、MJ:满减、N元购）
      */
     private String marketPlan;
+
     /**
      * 营销优惠表达式
      */
     private String marketExpr;
+
     /**
-     * 人群标签，特定限定优惠
+     * 人群标签，特定优惠限定
      */
     private String tagId;
+
     /**
      * 创建时间
      */
@@ -59,4 +67,5 @@ public class GroupBuyDiscount {
      * 更新时间
      */
     private Date updateTime;
+
 }
